flask_app/requirements.txt
flask>=2.3.0
flask-session>=0.5.0
pandas>=1.5.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
# For PostgreSQL support
psycopg2-binary>=2.9.5
# For MySQL support
pymysql>=1.0.2
cryptography>=36.0.0
# For LLM models
google-generativeai>=0.3.0
openai>=1.0.0
# For browser automation
playwright>=1.40.0 